src/main/java/com/example/sweater/web/UserController.java
package com.example.sweater.web;

import com.example.sweater.model.User;
import com.example.sweater.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
public class UserController {

    private final UserService userService;

    @Autowired
    public UserController(UserService userService) {
        this.userService = userService;
    }

    //for admin
    @PostMapping(value = "/users", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
    public User addNew(@RequestBody User user) {
        return userService.addNew(user);
    }

    //for admin
    @GetMapping(value = "/users/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
    public User getById(@PathVariable Integer id) {
        return userService.getById(id);
    }

    //for admin
    @GetMapping(value = "/users/by", produces = MediaType.APPLICATION_JSON_VALUE)
    public User getByEmail(@RequestParam(name = "email") String email) {
        return userService.getByEmail(email);
    }

    //for admin
    @GetMapping(value = "/users", produces = MediaType.APPLICATION_JSON_VALUE)
    public List<User> getAll() {
        return userService.getAll();
    }

    //for admin
    @PutMapping(value = "/users/{id}", produces = MediaType.APPLICATION_JSON_VALUE, consumes = MediaType.APPLICATION_JSON_VALUE)
    public void update(@PathVariable("id") Integer id,
                       @RequestBody User user) {
        userService.update(id, user);
    }

    //for admin
    @DeleteMapping(value = "/users/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
    public void deleteById(@PathVariable Integer id) {
        userService.deleteById(id);
    }
}
